# 7903: FastAccessNumberMap is slow for sparse values