@richardstartin 2022-09-29
> This commit violates spotless by not adding a line break at the end, see [here](https://github.com/SAP/jmc/actions/runs/3147570384/jobs/5117198800). I've opened a PR for the fix: #433
> 
> For the next time, please activate GHA in your jmc repository fork.

Oh no! I do apologise. I don't think I've ever had to enable GHA in a fork of any other project before in order to have CI run - are there any pointers on how to do this?