@RealCLanger 2022-09-29
Seems like you did it already. If I look [here](https://github.com/richardstartin/jmc/actions) I can see a run. But it was not testing the source commit of this PR.